Modified Files:  	configure.ac src/plugins/imexporters/Makefile.am  	src/plugins/imexporters/eri/eri.c  	src/plugins/imexporters/eri/eri_p.h
Hmm, thing behave certainly different, with editor set to
vim instead of gvim. Please let me know when this appears
in the Log.

Changed a comment in eri_p.h from C++ to standard C.
Not sure what changed in eri.c. I saved the stable version
and copied it back before commit. But cvs still wants to
commit it.


git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@392 5c42a225-8b10-0410-9873-89b7810ad06e