Added a CSV profile editor dialog.
The CSV profile editor now works and can be used to adjust existing profiles.
The generic importer dialog can use this new dialog now.
The function AB_ImExporter_GetEditProfileDialog now takes the name of a test
file instead of a GWEN_IO_LAYER. This is needed because the editor needs to
rewind the data when some parameters have changed, and GWEN_IO_LAYER can't do
that. Also, using a filename makes it much easier to use this function.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@1827 5c42a225-8b10-0410-9873-89b7810ad06e