Fixed a problem with Commerzbank.
Sending an RDH1 key using HBCI 2.20 and below requires the operation mode
of the key to be set to "16" even for the crypt key. This value is
actually wrong, and since FinTS 3.0 the value "2" is required (which is
the correct value for a crypt key). So we need to make that determination.

After this change the Commerzbank server accepts the keys.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@1923 5c42a225-8b10-0410-9873-89b7810ad06e