Added first version of a Paypal backend.
Please note the special license of this backend!

This backend can only be used to retrieve a list of transactions from
a paypal account. To use this you must apply for an API signature at
paypal (see homepage of paypal).

Since the Paypal backend needs some sensitive data which is to complex
to enter every time we connect to the server this backend uses the new
module GWEN_SmallTresor to store those secrets in a password protected
file.

This backend has not been extensively tested, so use on your own risk!
It is not compiled per default, you need to specify it with --with-backends.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@1940 5c42a225-8b10-0410-9873-89b7810ad06e