Fixed AH_User_GetTanMethodDescriptions().
This is quite a subtile bug: The list of TAN descriptions is normally read
when loading the user. However, this only happens when calling the
BeginExclUser function. The PinTan setup dialog doesn't do that between
mutliple tries of retrieving the system id, so although the list of tan
descriptions was retrieved it was never parsed into the list returned by
AH_User_GetTanMethodDescriptions between those tries. The result was the
inability to select a valid two-step tan method.
With this fix the list of TAN descriptions is always re-generated.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@1969 5c42a225-8b10-0410-9873-89b7810ad06e