Very much improved "new user" dialog.
Added a two-step dialog which selects the backend for which to create a
user and which in the second step allows selection of the user type to
create (e.g. pintan, keyfile, chipcard etc).

Online banking protocol modules now provide the widgets for the second
step. In total this change decreases the complexity of creating online
banking users.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2012 5c42a225-8b10-0410-9873-89b7810ad06e