Added support for sub accounts.
HBCI also nows about sub account ids ("Unterkontomerkmal") which is important
with some banks (namely Deutsch Bank). this particular bank provides
multiple accounts with the same bank code and account number but with a
different sub account id. So AqBanking must be able to differentiate between
those accounts. That's why handling of this id has been moved from AqHBCI
up to AqBanking.
AqBanking-CLI now uses the switch "-as" for this id so that such accounts
can now be selected.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2107 5c42a225-8b10-0410-9873-89b7810ad06e