AqHBCI-Tool4: Added command "logfile".
This command can be used to anonymize a logfile to send it for debugging
purposes.

The typical use of this command is this:
     "aqhbci-tool4 logfile -i LOGFILE -d OUTFILE -L 100"
"-L" specifies a very high trust level, meaning the output file contains
sensitive information. The default is "0" meaning no sensitive information
is written to the resulting file.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2200 5c42a225-8b10-0410-9873-89b7810ad06e