Reorganized code for current HKTAN segments.
This is really ugly... The ZKA decided to make the PIN/TAN stuff even more
complicated. It is nearly impossible to keep compatible with older HKTAN
segments and more recent ones.

This change now starts shifting the setup of challenge parameters from
HKTAN (where it really belongs) to the HBCI jobs. This is, as said above,
extremely ugly but necessary. And it matches the latest strategy of the ZKA.
On a side note: Of course some of the information needed to correctly setup
TAN challenge parameters is not publicly available...

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2265 5c42a225-8b10-0410-9873-89b7810ad06e