AqHBCI: Added a flag to omit SMS account specification.
Some banks choke on the SMS account specification in HKTAN. Normally, if
the bank doesn't need this information, it should just ignore it, but of
course this is not the case with some bank servers (sigh).
So now the user can set the new flag thereby making AqHBCI omit this
information.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2282 5c42a225-8b10-0410-9873-89b7810ad06e