Fixed SWIFT parser.
Some banks actually send transactions with a difference between booking
date and valuta date of more than 1 month. This is quiet problematic since
only the valuta date specifies a year but the booking date does not. So the
year of the booking date must be somehow derived from the valuta date.
We already adapted the algorithm to that effect some time ago, but a
difference of more than 1 month was not handled (who would have thought of
something like that).
Not having specified a booking year in SWIFT documents is really dumb...

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2283 5c42a225-8b10-0410-9873-89b7810ad06e