Fixed problems with OFX files.
OFX files are no normal XML files... Most of the elements are not properly
closed (i.e. there is no "</something>" element for "<something<" elements),
so when the "<OFX>" element is closed, we need to reset the XML context
depth, otherwise the GWEN_XmlCtx class won't correctly detect the end of
the OFX document.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2300 5c42a225-8b10-0410-9873-89b7810ad06e