AqBanking-CLI: Don't bother setting country and rcountry (Elias Oltmanns)
Except for chkacc, this never worked in the first place. None of the
AqBanking-CLI commands accepted the country (as in local country)
option, and the rcountry option, when specified, was ignored by
mkTransfer, hence had no effect. Since IBAN and BIC will make this
information redundant anyway, rather drop than fix the code. Leave
chkacc untouched since it actually works as advertised and may be
useful to someone out there.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2532 5c42a225-8b10-0410-9873-89b7810ad06e