SEPA exporter: Unify interface to pain.008.00x.0y documents
In an attempt to reduce code duplication in order to make it easier to
maintain, all functions creating pain.008 messages are merged into
one. Also, profiles are extended and shall make the code more
flexible, eventually.

The cor1 variants of the profiles below 008.003.02 are discarded since
the COR1 option has been defined in pain.008.003.02 for the first
time.

git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2565 5c42a225-8b10-0410-9873-89b7810ad06e