aqbanking-cli: Fix parameters for calling AH_Provider_SendDtazv from libaqhbci
The parameter 'withProgress' were added to function AH_Provider_SendDtazv in
SVN revision 1860. Prior to this change the function used a hard-coded 1
instead of the new parameter. So I figured that 1 is the correct value to
restore the behavior prior to rev. 1860. This fixes the compiler error
intentionally introduced by the previous commit.

This fixes a regression that was introduced in SVN revision 1860 and that was
discovered thanks to Michael Tautschnig <mt@debian.org> reporting Debian bug
#750046 ( http://bugs.debian.org/750046 ).


git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@2592 5c42a225-8b10-0410-9873-89b7810ad06e