Reworked SWIFT MT940 parser.
The Netbank/Sparda server is extremely nasty: The FinTS specs state, that
a new SEPA identifier must start in a new SWIFT subfield ("?20" etc).
However, this server just sends the SEPA multiple tags like "EREF+" et al
within a single subfield, so our parser, adhering to the specs, can't parse
the purpose string from this server...

The next step will have to be to just concatenate all purpose-related subfields
and try to read those SEPA tags from a single line. This is extremely ugly,
and it will also significantly slow down the parser, and all that because
those banking IT guys can't adhere to the specs which are extremely clean
regarding this point...