AqHBCI: "aqhbci-tool4 getsysid" first retrieves the bank cert.
Normally the certificate should be retrieved with
"aqhbci-tool4 getsysid".

However, if this step is omitted the user has to input the pin upon
"getsysid" before the certificate is presented to him for approval.

When explicitly retrieving the certificate here the user only has to
provide the pin after accepting the cert.

Anyway, the pin or other relevant data is (and always has been) in any
case only sent to the server *AFTER* the certificate has been accepted.