Fixed a AB_USER locking problem.
When creating/updating AB_ACCOUNT_SPEC objects, at least the backend
AqHBCI needs to access data from the corresponding AB_USER object.

When updating/writing an AB_ACCOUNT_SPEC upon message exchange with the
server the corresponding user already is locked, so lock requests must fail
and they do. So the function AB_Provider_AddAccount() needs to know whether
it is safe to try to lock the corresponding user or not.

This is just an internal change, no change to the API.