No longer complain about missing BICs.
BICs are no longer required in SEPA transactions. The BIC is also used
if present but a missing BIC should no longer be handled as an error.

Some bank servers might still require the BIC, in that case the user will
receive an error from the server.

However, the BIC is still checked for SEPA message versions prior to
pain.00x.003.02, as it is required by the the format definitions.