AqBanking: Added function AB_Transaction_GetPurposeAsStringList()
Previous versions of AqBanking used to store the purpose field as string list.
Current versions use a single string in which multiple lines are concatenated
by NEWLINE characters (#10).

To make transition easier this convenience function was added.

Please note that the caller is responsible for freeing the stringlist
returned (if any).