aqbanking-cli: Use GWEN_Text_ReplaceVars() instead of GWEN_DB_ReplaceVars().
This is a more generic function because it allows the use of a callback
to get the variable data. We can then define variables which use existing
variables (like "purpose") to create derivative data (e.g. "purposeInOneLine"
which removes CR characters in the purpose field and shows the whole
purpose in one line).

Needs latest GWEN.