OFX: Prepared complete move to XML-based importer also for OFX v1.
We can now safely read OFX v1 files with the XML importer, however, this
needed small changes to libgwenhywfar.

Now we just need a profile file for OFXv1 which I already started.
This is much easier than the old OFX v1 importer which will soon be removed.

Adding the ability to read new OFX structures to the old parser is extremely
painful, this is now much much easier with the generic XML importer (which
wasn't available when the old OFX parser was created).