aqhbci: started working on Verification of Payee.
DON'T USE THIS BRANCH IN PRODUCTION!!!

This branch represents work on implementing the newly required feature
"Verification of Payee" which requires the application to request a check
of the payee for every transaction.

This had automatically been done by the banks internally before SEPA.
With SEPA that feature was dropped leaving users open to fraud.

So for the convenience of the banks we now have to introduce yet another
layer around sending transactions to the bank.

This branch is not released to the general public because it still doesn't
handle "Verification of Payee", yet.

USE AT YOUR OWN RISK!

That being said, this code when finished shouldn't introduce systemic
problems since not verifying the payee is the default behaviour of
AqBankings master branch for now anyway.